% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminf4movenet-package.R
\docType{package}
\name{siminf4movenet-package}
\alias{siminf4movenet-package}
\alias{_PACKAGE}
\alias{siminf4movenet}
\title{What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
Some description of the package
}
\examples{
## An example function:
df <- example_function()
df

## To cite this package in publications use:
citation("siminf4movenet")

}
\references{
Some article that you might want users to look at
}
\seealso{
\code{\link[movenet]{movedata2networkDynamic}} for something in movenet
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Carlijn Bogaardt \email{Carlijn.Bogaardt@glasgow.ac.uk}
  \item Stefan Widgren \email{stefan.widgren@sva.se} (\href{https://orcid.org/0000-0001-5745-2284}{ORCID})
}

}
\keyword{internal}
