% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminf4movenet-package.R
\docType{package}
\name{siminf4movenet-package}
\alias{siminf4movenet-package}
\alias{_PACKAGE}
\alias{siminf4movenet}
\title{What the Package Does (One Line, Title Case)}
\description{
A `SimInf` model designed for the `movenet` package, which facilitates the use of livestock movement data in veterinary public health. The SEIRcm model is a holding-unit SEIR model with between-holding infectious pressures supplied via a contact matrix.
}
\details{
Some description of the package
\subsection{To cite this package in publications use:}{

citation("siminf4movenet")
}
}
\references{
Some article that you might want users to look at
}
\seealso{
\code{\link[movenet]{movedata2networkDynamic}} for something in movenet
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Carlijn Bogaardt \email{Carlijn.Bogaardt@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-9322-6978}{ORCID})
  \item Jess Enright \email{Jess.Enright@glasgow.ac.uk}
  \item Stefan Widgren \email{stefan.widgren@sva.se} (\href{https://orcid.org/0000-0001-5745-2284}{ORCID})
}

}
\keyword{internal}
